requirements.txt
gunicorn==19.9.0
pandas==1.1.3
google-cloud-vision==2.3.1
Flask==1.1.
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
requests==2.22.0
Werkzeug==0.16.0
